Add experiment logging
Add global .scientist.setLogger and use as part of every dynamically
created function returned from 'new'.  If we used experimental
function as part of calling scientist, check the values match and
report to logger, including differences in values in the logged
string.

.scientist.new now returns a dictionary containing experiment ID and
function to use.  This makes it simpler to write tests for logger
functionality.

Note that logger functionality is not enabled by default but should be
set after loading code.  It might be a good idea to make this specific
to each function, or to allow functions to override the default going
forward, but we have the flexibility to do this later.

As we are now getting to the point where multiple tests use the same
setup, collapse these into a single functions, beforesimple and
cleanup.  Due to the issues with using functions defined in the .q
namespace, we need to perform a little trickery on the beforesimple
function, wrapping it up as a parameter to a function which
re-evaluates the function after qspec functions have been defined.