6 - Replace createExperiment with composition implementation
Code no longer create functions dynamically from strings in
.scientist.new.  Instead, use composition and varadic enlist to
collect arguments into a list containing three values:

  - a sentinel.  This is there to prevent unintended list collapse,
    should we decide to change the way we do indexing into
    .scientist.try

  - index of functions to use for this experiment in .scientist.try.

  - arguments to experiment functions.

This takes advantage of two special bits of syntax in q:

  - enlist can be used in the following manner:

    q)enlist[1;2;3;4;5;6;7;8;9]
    1 2 3 4 5 6 7 8 9

  - Lists can have arguments elided and used as functions, allowing us
    to create the three-item list.

createExperiment then uses a layer of indirection, calling internal
function i.createExperiment which separates out the various parts
above on calling the experiment, and calls the use and try functions
as before.

This also vastly simplifies the code in .scientist.new, which no
longer needs to count args, make strings of index arguments, or pass
arguments to createExperiment.

This necessitated a slight change in the logger string formatting, as
arguments are now presented to it as a list even for single arguments.
It's possible that displaying arguments in list format is more
correct, so for now let's just change the expected format.