10 - Call default error handler
If a user does not specify the error handler on per-experiment basis,
call a default error handler instead.

If either 'use' or 'try' function throws, error handler is called.

This requires the experimentRunner code to trap any errors in the 'use'
function, call the error handler, and then re-throw the error.