17 - only call beforeRun if experiment enabled
Extend existing test to check logging outputs for enabled and disabled
tests.  If disabled, only enabler callback should be called for init;
beforeRun should not be called as we are intending to avoid expensive
operations.

If experiment is enabled, we expect to see enabler called at init and
preExperiment, and then the beforeRun function called following that.

This was mainly done by swapping the order of enabler and beforeRun
calls in experimentRunner.