22 - Allow users to provide experiment names
Add tests:

  - should allow user to pass a name for experiments

  - should return internal reference as name of ewxperiment if none
    provided

Note that we have not specified that experiment name should not be of
uniform type across experiments.  No idea what this will mean though.
If it becomes a problem we can change the behvaiour later.

In particular if we later switch to an internal reference that is not
atomic, what will this do?  Possible source of unintended coupling
here...

Also refactored some other tests that were over-specific in what return
values they accepted from .scientist.new